requirements.txt
fastapi~=0.101.1
SQLAlchemy~=2.0.19
pydantic~=2.1.1
python-dotenv~=1.0.0
pytest~=7.4.0
psycopg2-binary~=2.9.7
uvicorn~=0.23.2